Enable custom debugger that stores XMPP output per test in different files
When tests fail, it can be hard to determine the cause of the failure. This commit enables a custom Smack debugger that will store the XMPP traffic for every test in an output file that is specific to that test. These files are all stored in a directory that is configurable using the `logDir` system property.