Do not generate Github 'check' for JUnit report
Our Github Action should be the only 'check' in a PR that utilizes it (a 'check' being the one item in the list of things that get green marks or red crosses whenever a PR is being built).

The JUnit Report action that's embedded in _our_ action should not generate a check of its own, as it's an integral part of our action. Additionally, to create a check, it apprently requires to be given permission in the workflow that uses it, which is another reason to not want it: it would add an additional configuration requirement for users of our action.