Reduce default time-out value
Tests are expected to finish quickly. When a timeout occurs, this is pretty much always a sign of a test failure.

As tests are executed in a test environement, in close proximity to the server under test, the default timeout can be set quite low: there is no need to account for network congestion.

Having a low default timeout value allows the failure reporting to be a lot quicker.

If need be, individual users can use the configuration option to increase the timeout value again.