Use SINT 'host' configuration option
Recently, the Smack Integration Test Framework got a new configuration option, named 'host'. This allows the test to be configured with an IP address or DNS name of the XMPP service to run the tests on.

This new option negates the tricks involving setting `/etc/hosts` to make the XMPP domain name resolve to the intended host. This commit replaces that solution with the new configuration option.

Additionally, this commit renames the old 'ipAddress' configuration option to 'host', to improve consistency in nomenclature.