Add input to differentiate between executions
When the Action is used in more than one job for a workflow, there is a need to differentiate between the different executions. For one, the output that is uploaded for each execution needs to have a unique name (as they can't overwrite each-other).

This commit adds an input named `testId` that can be used for this purpose.

fixes #37