Allow tests to be added via a new classpath argument
This adds a new optional configuration option (`extensionJar`) that can be used to add single JAR file to the exection, to load custom test extensions or additional test logic.

When there is an additional jar file for the classpath, it is merged into the main jar. I've tried to instead add the additional data to the classpath, but that didn't work (although I still believe that that would be the right approach).

fixes #52